Master Password info added to file->properties dialog
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/branches/mpwpol@2143 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ corelib/PWSfile.h
@@ -126,6 +126,6 @@ protected:
   // Save unknown header fields on read to put back on write unchanged
   UnknownFieldList m_UHFL;
   int m_nRecordsWithUnknownFields;
-  long m_fileLength;
+  size_t m_fileLength;
 };
 #endif /* __PWSFILE_H */
